requirements.txt
duckdb==0.9.2
python-dotenv==1.0.0
